requirements.txt
Django~=3.2
django-cleanup~=5.2.0
django-environ~=0.7.0
django-markdown-deux~=1.0.5
django-meta~=2.0.0
django-pagedown~=2.2.1
markdown2~=2.4.1
psycopg2-binary~=2.9.1
python-telegram-bot~=13.7
requests~=2.26.0
schedule~=1.1.0
stripe~=2.60.0
uWSGI~=2.0.19
wheel~=0.37.0
